+++ src/main/java/org/uom/lefterisxris/codetour/tours/service/Navigator.java
@@ -39,10 +39,11 @@ public class Navigator {
          }
 
          // Navigate
-         new OpenFileDescriptor(project, virtualFile.get(), step.getLine(), 1).navigate(true);
+         new OpenFileDescriptor(project, virtualFile.get(), Math.max(step.getLine() - 1, 0), 1)
+               .navigate(true);
 
          // Show a Balloon
-         CodeTourNotifier.notifyStepDescription(project, step.getDescription());
+         // CodeTourNotifier.notifyStepDescription(project, step.getDescription());
 
          // Show a Popup
          //TODO: It would be nice to provide the Tour as well, for info like 1/5 steps etc