+++ src/main/java/org/uom/lefterisxris/codetour/tours/ui/ToolPaneWindow.java
@@ -2,6 +2,7 @@ package org.uom.lefterisxris.codetour.tours.ui;
 
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.fileEditor.OpenFileDescriptor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.JBPopupMenu;
 import com.intellij.openapi.ui.Messages;
@@ -12,11 +13,14 @@ import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.ui.components.JBScrollPane;
 import com.intellij.ui.treeStructure.Tree;
 import com.intellij.util.SlowOperations;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.uom.lefterisxris.codetour.tours.domain.Props;
 import org.uom.lefterisxris.codetour.tours.domain.Step;
 import org.uom.lefterisxris.codetour.tours.domain.Tour;
 import org.uom.lefterisxris.codetour.tours.service.Navigator;
+import org.uom.lefterisxris.codetour.tours.service.TourValidator;
+import org.uom.lefterisxris.codetour.tours.service.Utils;
 import org.uom.lefterisxris.codetour.tours.state.StateManager;
 import org.uom.lefterisxris.codetour.tours.state.StepSelectionNotifier;
 import org.uom.lefterisxris.codetour.tours.state.TourUpdateNotifier;
@@ -29,7 +33,7 @@ import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.util.List;
 import java.util.*;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
 
 /**
  * Code Tour - Tool Window (Tours Navigation and Management).
@@ -41,17 +45,20 @@ import java.util.concurrent.atomic.AtomicBoolean;
  */
 public class ToolPaneWindow {
 
+   private static final String ID = "Tours Navigation";
    private static final Logger LOG = Logger.getInstance(ToolPaneWindow.class);
    private static final String TREE_TITLE = "Code Tours";
 
    private final JPanel panel;
    private Tree toursTree;
 
+   private final ToolWindow toolWindow;
    private final Project project;
    private final StateManager stateManager;
 
    public ToolPaneWindow(@NotNull Project project, @NotNull ToolWindow toolWindow) {
 
+      this.toolWindow = toolWindow;
       this.project = project;
       this.stateManager = new StateManager(project);
       panel = new JPanel(new BorderLayout());
@@ -79,7 +86,11 @@ public class ToolPaneWindow {
       });
 
       project.getMessageBus().connect().subscribe(StepSelectionNotifier.TOPIC, (step) -> {
-         StateManager.getActiveTour().ifPresent(tour -> selectTourStep(tour, StateManager.getActiveStepIndex()));
+         StateManager.getActiveTour().ifPresent(tour -> {
+            if (!toolWindow.isVisible())
+               toolWindow.show();
+            selectTourStep(tour, StateManager.getActiveStepIndex());
+         });
       });
    }
 
@@ -144,7 +155,7 @@ public class ToolPaneWindow {
 
    private void createNavigationButtons() {
       final JButton previousButton = new JButton("Previous Step");
-      previousButton.setToolTipText("Navigate to the Previous Step of the active Tour");
+      previousButton.setToolTipText("Navigate to the Previous Step of the active Tour (Ctrl+Alt+Q)");
       previousButton.addActionListener(e -> {
          LOG.info("Previous button pressed!");
 
@@ -156,7 +167,7 @@ public class ToolPaneWindow {
       });
 
       final JButton nextButton = new JButton("Next Step");
-      nextButton.setToolTipText("Navigate to the Next Step of the active Tour");
+      nextButton.setToolTipText("Navigate to the Next Step of the active Tour (Ctrl+Alt+W)");
       nextButton.addActionListener(e -> {
          LOG.info("Next button pressed!");
 
@@ -211,11 +222,15 @@ public class ToolPaneWindow {
          final JMenuItem editAction = new JMenuItem("Edit Tour", AllIcons.Actions.Edit);
          editAction.addActionListener(d -> editTourListener(tour));
 
+         // Jump to Source Action
+         final JMenuItem jumpToSourceAction = new JMenuItem("Jump to .tour Source", AllIcons.Actions.EditSource);
+         jumpToSourceAction.addActionListener(d -> jumpToSourceTourListener(tour));
+
          // Delete Action
          final JMenuItem deleteAction = new JMenuItem("Delete Tour", AllIcons.Actions.DeleteTag);
          deleteAction.addActionListener(d -> deleteTourListener(tour));
 
-         Arrays.asList(newStepAction, editAction, deleteAction).forEach(menu::add);
+         Arrays.asList(newStepAction, editAction, jumpToSourceAction, deleteAction).forEach(menu::add);
          menu.show(toursTree, e.getX(), e.getY());
       }
    }
@@ -255,33 +270,37 @@ public class ToolPaneWindow {
    //endregion
 
    private void createNewTourListener() {
-      final long index = stateManager.getTours().stream()
-            .map(tour -> tour.getTourFile())
-            .filter(tourFile -> tourFile != null)
-            .filter(tourFile -> tourFile.startsWith("newTour") && tourFile.endsWith(Props.TOUR_EXTENSION_FULL))
-            .count();
-      String newTourFileName = String.format("newTour%s.%s", (index > 0 ? index : ""), Props.TOUR_EXTENSION);
-
       final Tour newTour = Tour.builder()
             .id(UUID.randomUUID().toString())
-            .touFile(newTourFileName)
+            .touFile("newTour" + Props.TOUR_EXTENSION_FULL)
             .title("A New Tour")
             .description("A New Tour")
             .steps(new ArrayList<>())
             .build();
 
+      // Interactive creation (Title and filename) making sure that they are unique
+      final Set<String> tourTitles = stateManager.getTours().stream()
+            .map(Tour::getTitle)
+            .collect(Collectors.toSet());
+      final String updatedTitle = Messages.showInputDialog(project,
+            "Input the title of the new Tour (should be unique)",
+            "New Tour", AllIcons.Actions.NewFolder, newTour.getTitle(),
+            new TourValidator(title -> StringUtils.isNotEmpty(title) && !tourTitles.contains(title)));
+      if (updatedTitle == null) return; // i.e. hit cancel
+      newTour.setTitle(updatedTitle);
+
+
       // Just make sure that the new file is unique
-      AtomicBoolean nameIsFine = new AtomicBoolean(false);
-      int nTry = 1;
-      while (!nameIsFine.get()) {
-         SlowOperations.allowSlowOperations(() -> {
-            final Collection<VirtualFile> files =
-                  FilenameIndex.getVirtualFilesByName(newTour.getTourFile(), GlobalSearchScope.projectScope(project));
-            if (files.isEmpty()) nameIsFine.set(true);
-            else newTour.setTourFile(
-                  String.format("newTour%s-%s.%s", (index > 0 ? index : ""), nTry, Props.TOUR_EXTENSION));
-         });
-      }
+      final Set<String> tourFiles = stateManager.getTours().stream()
+            .map(Tour::getTourFile)
+            .collect(Collectors.toSet());
+      final String updatedFilename = Messages.showInputDialog(project,
+            "Input the file name of the new Tour (should end with .tour and be unique)",
+            "New Tour", AllIcons.Actions.NewFolder, Utils.fileNameFromTitle(newTour.getTitle()),
+            new TourValidator(fileName -> StringUtils.isNotEmpty(fileName) &&
+                  fileName.endsWith(Props.TOUR_EXTENSION_FULL) && !tourFiles.contains(fileName)));
+      if (updatedFilename == null) return; // i.e. hit cancel
+      newTour.setTourFile(updatedFilename);
 
       stateManager.createTour(newTour);
       createToursTee(project);
@@ -291,24 +310,11 @@ public class ToolPaneWindow {
 
    //region Tour Context menu actions
    private void addNewStepOnTourListener(Tour tour) {
-      final long index = tour.getSteps().stream()
-            .filter(step -> step.getTitle().startsWith("New Step"))
-            .count();
-
-      final String stepTitle = String.format("New Step%s", (index > 0) ? index : "");
-      tour.getSteps().add(Step.builder()
-            .title(stepTitle)
-            .description("Step Description")
-            .file("")
-            .line(1)
-            .build());
-      stateManager.updateTour(tour);
-      createToursTee(project);
-      CodeTourNotifier.notifyTourAction(project, tour, "New Step",
-            String.format("A New Step added on Tour '%s'", tour.getTitle()));
-
-      // Expand and select the last Step of the active Tour on the tree
-      selectTourLastStep(tour);
+      Messages.showMessageDialog(project,
+            "To create a new Tour Step, Navigate to the file you want to add a Step, " +
+                  "<kbd>Right Click on the Editor's Gutter</kbd> (i.e. next to line numbers) > <kbd>Add Tour Step</kbd>",
+            "Step Creation",
+            AllIcons.General.NotificationInfo);
    }
 
    private void editTourListener(Tour tour) {
@@ -328,6 +334,24 @@ public class ToolPaneWindow {
 
       // Expand and select the last Step of the active Tour on the tree
       selectTourLastStep(tour);
+   }
+
+   private void jumpToSourceTourListener(Tour tour) {
+      SlowOperations.allowSlowOperations(() -> {
+         final Collection<VirtualFile> virtualFiles = FilenameIndex.getVirtualFilesByName(tour.getTourFile(),
+               GlobalSearchScope.projectScope(project));
+         final Optional<VirtualFile> virtualFile = virtualFiles.stream()
+               .filter(file -> !file.isDirectory() && file.getName().equals(tour.getTourFile()))
+               .findFirst();
+         if (virtualFile.isEmpty()) {
+            CodeTourNotifier.error(project, String.format("Could not locate navigation target '%s' for Tour '%s'",
+                  tour.getTourFile(), tour.getTitle()));
+            return;
+         }
+
+         // Navigate
+         new OpenFileDescriptor(project, virtualFile.get(), 0).navigate(true);
+      });
    }
 
    private void deleteTourListener(Tour tour) {
@@ -363,8 +387,11 @@ public class ToolPaneWindow {
    }
 
    private void editStepDescriptionListener(Step step, Tour tour) {
-      final String updatedDescription = Messages.showMultilineInputDialog(project, "Edit Step's description",
-            "Edit Step", step.getDescription(), AllIcons.Actions.Edit, null);
+      final String updatedDescription = Messages.showMultilineInputDialog(project,
+            String.format("Edit Step's '%s' description", step.getTitle()),
+            "Edit Step",
+            step.getDescription(),
+            AllIcons.Actions.Edit, null);
       if (updatedDescription == null || updatedDescription.equals(step.getDescription())) return;
 
       final Optional<Step> origStep = tour.getSteps().stream()