+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # CodeTour Changelog
 
 ## [Unreleased]
-### Added
+### Implemented
 - Load from json, Show In Tool Pane, Navigate with click
 - Navigate with Previous/Next buttons on Tool Pane
 - Create|Delete Tour from Tool Pane
@@ -11,3 +11,7 @@
 - Edit steps (CRUD + add text)
 - Preview Step description on the code (as a resizable/movable modal, after Navigation)
 - HTML support on steps
+- Persisting user preferences on Size/Location of Step Documentation modal
+- Gutter Icons for Steps on code
+- Jump to Tour source file (via Tree click)
+- Steps navigation through Step Documentation modal (arrow buttons)